GEN-1457 | Move partner service to widget feature
