+++ apps/store/src/features/widget/publishWidgetEvent.ts
@@ -1,17 +1,17 @@
-export type PartnerEvent =
+export type WidgetEvent =
   | { status: 'success' }
   | { status: 'cancel' }
   | { status: 'error' }
   | { status: 'close' }
   | { status: 'event'; message: { requestId: string } }
 
-export const publishPartnerEvent = (event: PartnerEvent): void => {
+export const publishWidgetEvent = (event: WidgetEvent): void => {
   if (window.ReactNativeWebView) {
     window.ReactNativeWebView.postMessage(JSON.stringify(event))
-    console.debug(`Partner event (native) sent:`, event)
+    console.debug(`Widget event (native) sent:`, event)
     return
   }
 
   window.parent.postMessage(event, '*')
-  console.debug(`Partner event (web) sent:`, event)
+  console.debug(`Widget event (web) sent:`, event)
 }
